requirements.txt
Django>=4.2,<4.3
djangorestframework>=3.14.0
djangorestframework-simplejwt>=5.3.0
django-cors-headers>=4.3.0
PyJWT==2.8.0
drf-yasg>=1.21.0
python-dotenv>=1.0.0
